requirements.txt
discord.py
cohere
langchain
langchain-community
langchain-cohere
faiss-cpu
flask
python-dotenv
tiktoken
